third_party/WebKit/LayoutTests/fast/dom/event-attrs-isolated-world-expected.txt

CONSOLE MESSAGE: line 1: PASS: Main world object not visible in isolated world
CONSOLE MESSAGE: line 1: PASS: Got back isolated world wrapper
CONSOLE MESSAGE: line 1: PASS: Declarative onclick handler did not leak to the isolated world
CONSOLE MESSAGE: line 27: PASS: Got back main world wrapper
DOM wrappers should not leak across world boundaries via onevent attributes
